Primer test de abondonar en verde pero con varias colaboradores ignoradas

+++ src/main/java/ar/edu/unq/epers/bichomon/backend/model/duelo/ResultadoCombate.java
@@ -24,7 +24,7 @@ public class ResultadoCombate {
     public String getNombreDeEntrenadorCampeon(){
         return entrenadorCampeon.getNombre();
     }
-    public Long getBichoCampeonId(){ return bichoCampeon.getId();}
+    public Integer getBichoCampeonId(){ return bichoCampeon.getId();}
     public List getAtaques(){
         return ataques;
     }
